requirements.txt
streamlit
langchain
langchain_groq
langchain_community
PyPDF2
tempfile
pydantic
python-dotenv
